Added pre_tax_price and pre_tax_price_set properties
Both the pre_tax_price and pre_tax_price_set properties are available on most Shopify Plus shops, but this is not (publicly) documented. By adding this it becomes possible to test and use the value without the need to calculate pre_tax_price values for products. Very useful in customs applications...